requirements.txt
aiogram>=3.4.0
python-dotenv>=1.0.0
yt-dlp>=2024.11.4
aiofiles>=24.1.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
pre-commit>=3.7.0
